add manual.md
