+++ luce_vm/scripts/simulation/Manual.md
@@ -1,0 +1,80 @@
+## Developing and Integrating New Simulation Strategies
+
+### 1. Introduction
+
+The Simulator system has been designed to be flexible and extensible. The use of the Strategy Pattern allows for the easy addition of new simulation scenarios without altering the core simulator functionality. This document provides a step-by-step guide to develop new strategies and integrate them into the existing system.
+
+### 2. Developing a New Strategy
+
+#### 2.1. Create a New Strategy Class
+
+Each new strategy should be a class in the `simulation_strategy.py` file. The class should inherit from the base `SimulationStrategy` class.
+
+```python
+from abc import ABC, abstractmethod
+
+class SimulationStrategy(ABC):
+
+    @abstractmethod
+    def execute(self, simulator):
+        pass
+```
+
+#### 2.2. Implement the Strategy
+
+Override the `execute` method in your new strategy class. This method contains the main logic for the simulation scenario.
+
+For example:
+
+```python
+class NewSimulationStrategy(SimulationStrategy):
+
+    def execute(self, simulator):
+        # Your new strategy's logic here
+        pass
+```
+
+### 3. Integrating the New Strategy with the Simulator
+
+#### 3.1. Import the Strategy
+
+In `simulator.py`, import your new strategy:
+
+```python
+from simulation_strategy import NewSimulationStrategy
+```
+
+#### 3.2. Use the Strategy
+
+When creating a new `Simulator` object, you can set your new strategy as follows:
+
+```python
+strategy = NewSimulationStrategy()
+simulator = Simulator(strategy)
+```
+
+Alternatively, if a `Simulator` object already exists, you can change its strategy:
+
+```python
+simulator.set_strategy(NewSimulationStrategy())
+```
+
+### 4. Running the New Strategy
+
+Once the strategy is set, simply run the simulator:
+
+```python
+simulator.run()
+```
+
+This will execute the logic defined in your `NewSimulationStrategy` class.
+
+### 5. Tips for Developing Strategies
+
+- **Encapsulation**: Each strategy should be self-contained. It should have all the necessary logic to execute the simulation scenario without relying on external state.
+- **Reusability**: Consider reusing existing methods in the `Simulator` class to avoid code duplication. For example, methods like `_login`, `_register`, etc. can be used across multiple strategies.
+
+
+### 6. Conclusion
+
+The modular design of the Simulator system makes it easy to extend with new simulation scenarios. By following the steps outlined in this guide, developers can seamlessly integrate new strategies and enrich the capabilities of the simulator.
