+++ luce_vm/scripts/simulation/simulator.py
@@ -1,26 +1,34 @@
-import urllib3
-import json
-
+# Standard libraries
 import sys
 import os
+import json
+import urllib3
+from abc import ABC, abstractmethod
+
+# Third-party libraries
 import django
-
-from generate_user import generate_users
-
-luce_django_path = os.path.abspath('../..')  # luce_vm
-luce_django_path = os.path.join(luce_django_path, 'luce_django/luce')
-sys.path.insert(0, luce_django_path)
-os.environ['DJANGO_SETTINGS_MODULE'] = 'lucehome.settings'
-django.setup()
-
-from accounts.models import User
-
 import networkx as nx
 import matplotlib.pyplot as plt
 
+# Constants
+LUCE_DJANGO_PATH = os.path.abspath(os.path.join('../..', 'luce_django/luce'))
+DJANGO_SETTINGS_MODULE = 'lucehome.settings'
+
+
+def setup_django():
+    sys.path.insert(0, LUCE_DJANGO_PATH)
+    os.environ['DJANGO_SETTINGS_MODULE'] = DJANGO_SETTINGS_MODULE
+    django.setup()
+
+
+setup_django()
+# Local modules
+from generate_user import generate_users
+from accounts.models import User
+
 
 class Simulator:
-    def __init__(self, num_of_users):
+    def __init__(self, num_of_users, strategy=None):
         self.http = urllib3.PoolManager()
         self.urls = {
             "login": "http://127.0.0.1:8000/user/login/",
@@ -32,10 +40,16 @@ class Simulator:
         self.user = generate_users(num_of_users)
         self.directed_graph = nx.DiGraph()
 
-        self.switcher = {1: self._senario_1}
+        self.strategy = strategy
 
-    def run(self, senario_num):
-        self.switcher.get(senario_num, lambda: "Invalid senario")()
+    def set_strategy(self, strategy):
+        self.strategy = strategy
+
+    def run(self):
+        if self.strategy:
+            self.strategy.execute(self)
+        else:
+            print("No strategy set.")
 
     def _login(self, url, user):
         encoded_user = json.dumps(user).encode('utf-8')
@@ -47,6 +61,12 @@ class Simulator:
         token = json.loads(r.data.decode('utf-8'))["data"]["token"]
         return token
 
+    def register_users(self):
+        for user in self.user['users']:
+            email = user['registration_data']['email']
+            print("Register user: " + email)
+            self._register(self.urls['register'], user['registration_data'])
+
     def _register(self, registration_url, registration_data):
         r = self.http.request(
             'POST',
@@ -55,8 +75,25 @@ class Simulator:
             headers={'Content-Type': 'application/json'})
 
         result = json.loads(r.data.decode('utf-8'))
-        # print(result)
-        return result
+
+    def upload_data(self):
+        for user in self.user['users']:
+            email = user['registration_data']['email']
+            print("Upload data for user: " + email)
+            token = self._login(
+                self.urls['login'], {
+                    'username': email,
+                    'password': user['registration_data']['password']
+                })
+            user_instance = User.objects.get(email=email)
+            user_address = user_instance.ethereum_public_key
+
+            uploaded = self._upload_data(self.urls['upload_data'],
+                                         user['uploaded_data'], token)
+            self.directed_graph.add_edge(
+                self._address_to_label(user_address),
+                self._address_to_label(
+                    uploaded['data']['contracts']["contract_address"]))
 
     def _upload_data(self, upload_url, data, token):
         d = json.dumps(data).encode('utf-8')
@@ -72,7 +109,16 @@ class Simulator:
         result = json.loads(r.data.decode('utf-8'))
         return result
 
-    def _deploy_registry(self, url, admin_token):
+    def deploy_registry(self):
+        admin = self.user['users'].pop()
+        admin_token = self._login(
+            self.urls['login'], {
+                'username': admin['registration_data']['email'],
+                'password': admin['registration_data']['password']
+            })
+        self._deploy_registry(admin_token)
+
+    def _deploy_registry(self, admin_token):
         print("Deploy registry")
         r = self.http.request("POST",
                               self.urls['deploy_registry'],
@@ -85,58 +131,9 @@ class Simulator:
     def _address_to_label(self, address):
         return address[:5] + "..." + address[-5:]
 
-    def _senario_1(self):
-        """
-        In this senario, we will simulate uploading data.
-        """
-        print("Start simulate senario 1")
-
-        # 1. clear user data
+    def _clear_data(self):
         User.objects.all().delete()
 
-        # 2. register users
-        for user in self.user['users']:
-            email = user['registration_data']['email']
-            print("Register user: " + email)
-            self._register(self.urls['register'], user['registration_data'])
-
-        # 3. deploy registry
-        admin = self.user['users'].pop()
-        admin_token = self._login(
-            self.urls['login'], {
-                'username': admin['registration_data']['email'],
-                'password': admin['registration_data']['password']
-            })
-        self._deploy_registry(self.urls['deploy_registry'], admin_token)
-
-        # 4. upload data
-        print("Upload data:")
-        for user in self.user['users']:
-            # 4.1 login
-            token = self._login(
-                self.urls['login'], {
-                    'username': user['registration_data']['email'],
-                    'password': user['registration_data']['password']
-                })
-
-            email = user['registration_data']['email']
-            user_instance = User.objects.get(email=email)
-            user_address = user_instance.ethereum_public_key
-
-            # 4.2 upload data
-            uploaded = self._upload_data(self.urls['upload_data'],
-                                         user['uploaded_data'], token)
-            print("Upload data for user: " + email)
-            uploaded_address = uploaded['data']['contracts']
-            print("Uploaded address: " + uploaded_address['contract_address'])
-            print()
-            self.directed_graph.add_edge(
-                self._address_to_label(user_address),
-                self._address_to_label(uploaded_address['contract_address']))
-
+    def _draw_graph(self):
         nx.draw(self.directed_graph, with_labels=True)
         plt.show()
-
-
-s = Simulator(3)
-s.run(1)
